refactor of term code & removed server for backend
