--- client/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "client-test",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "react-router build",
-    "dev": "react-router dev",
-    "start": "PORT=2705 react-router-serve ./build/server/index.js"
-  },
-  "dependencies": {
-    "@ant-design/icons": "^5.6.1",
-    "@ant-design/v5-patch-for-react-19": "^1.0.3",
-    "@react-router/node": "^7.5.3",
-    "@react-router/serve": "^7.5.3",
-    "@remix-run/node": "^2.16.7",
-    "@xterm/addon-fit": "^0.10.0",
-    "@xterm/xterm": "^5.4.0",
-    "antd": "^5.25.1",
-    "daisyui": "^5.0.40",
-    "isbot": "^5.1.27",
-    "js-cookie": "^3.0.5",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "react-router": "^7.5.3",
-    "remix-utils": "^8.7.0",
-    "zod": "^3.25.30"
-  },
-  "devDependencies": {
-    "@react-router/dev": "^7.5.3",
-    "@tailwindcss/vite": "^4.1.4",
-    "tailwindcss": "^4.1.4",
-    "vite": "^6.3.3"
-  },
-  "proxy": "http://localhost:1711"
-}
