--- server/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "body-parser": "^2.2.0",
-    "cors": "^2.8.5",
-    "crypto-js": "^4.2.0",
-    "express": "^5.1.0",
-    "multer": "^1.4.5-lts.2"
-  }
-}
